+++ requirements.txt
@@ -3,7 +3,9 @@ certifi==2021.10.8
 charset-normalizer==2.0.12
 click==8.0.3
 colorama==0.4.4
+dnspython==2.2.0
 docopt==0.6.2
+email-validator==1.1.3
 Flask==2.0.3
 Flask-Session==0.4.0
 Flask-WTF==1.0.0
