+++ urls4irl/forms.py
@@ -8,7 +8,7 @@ URL Creation form?
 """
 
 from flask_wtf import FlaskForm
-from wtforms import StringField, PasswordField, SubmitField
+from wtforms import StringField, PasswordField, SubmitField, EmailField
 from wtforms.validators import DataRequired, Length, Email, EqualTo
 
 
@@ -25,7 +25,7 @@ class UserRegistrationForm(FlaskForm):
     """
 
     username = StringField('Username', validators=[DataRequired(), Length(min=2, max=20)])
-    email = StringField('Email', validators=[DataRequired(), Email()])
+    email = EmailField('Email', validators=[DataRequired(), Email()])
     confirm_email = StringField('Confirm Email', validators=[DataRequired(), EqualTo(email)])
     password = PasswordField('Password', validators=[DataRequired()])
     confirm_password = PasswordField('Password', validators=[DataRequired(), EqualTo(password)])
